数据库表的id 类型由Integer改为Long

+++ src/main/java/life/majiang/community/community/controller/QuestionController.java
@@ -15,7 +15,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String  queation(@PathVariable(name = "id") Integer id
+    public String  queation(@PathVariable(name = "id") Long id
     , Model model)
     {
